# .agent/tmp/robyn_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OAP LangGraph Runtime",
    "version": "0.1.0",
    "description": "\nRobyn-based LangGraph-compatible runtime for Open Agent Platform.\n\nThis API provides endpoints for managing assistants, threads, runs, and the store.\nIt implements the LangGraph API specification for compatibility with LangGraph clients.\n"
  },
...
  "paths": {
...
    "/threads/{thread_id}/runs": {
      "get": {
        "tags": ["Thread Runs"],
        "summary": "List Thread Runs",
        "description": "List all runs for a thread.",
        "operationId": "list_thread_runs",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "The ID of the thread."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 10 }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "List of runs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Run" }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Thread Runs"],
        "summary": "Create Thread Run",
        "description": "Create a new run on a thread (background execution).",
        "operationId": "create_thread_run",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "The ID of the thread."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RunCreateStateful" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Run created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Run" }
              }
            }
          },
          "409": {
            "description": "Conflict - Concurrent run rejected",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/threads/{thread_id}/runs/wait": {
      "post": {
        "tags": ["Thread Runs"],
        "summary": "Wait for Thread Run",
        "description": "Create a run and wait for completion, returning final output.",
        "operationId": "wait_thread_run",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "The ID of the thread."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RunCreateStateful" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Final run output",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/threads/{thread_id}/runs/{run_id}": {
      "get": {
        "tags": ["Thread Runs"],
        "summary": "Get Run",
        "description": "Get a specific run by ID.",
        "operationId": "get_run",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "The ID of the thread."
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "The ID of the run."
          }
        ],
        "responses": {
          "200": {
            "description": "The requested run",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Run" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Thread Runs"],
        "summary": "Delete Run",
        "description": "Delete a run by ID.",
        "operationId": "delete_run",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "The ID of the thread."
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "The ID of the run."
          }
        ],
        "responses": {
          "200": {
            "description": "Run deleted successfully",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/threads/{thread_id}/runs/{run_id}/cancel": {
      "post": {
        "tags": ["Thread Runs"],
        "summary": "Cancel Run",
        "description": "Cancel a running execution.",
        "operationId": "cancel_run",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "The ID of the thread."
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "The ID of the run."
          }
        ],
        "responses": {
          "200": {
            "description": "Run cancelled",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["detail"],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Error message describing what went wrong."
          }
        },
        "title": "ErrorResponse",
        "description": "Standard error response format."
      },
...
      "Run": {
        "type": "object",
        "required": [
          "run_id",
          "thread_id",
          "assistant_id",
          "created_at",
          "updated_at",
          "status",
          "metadata"
        ],
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run ID",
            "description": "The unique identifier of the run."
          },
          "thread_id": {
            "type": "string",
            "format": "uuid",
            "title": "Thread ID",
            "description": "The ID of the thread this run belongs to."
          },
          "assistant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assistant ID",
            "description": "The ID of the assistant used for this run."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "The time the run was created."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "The last time the run was updated."
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "success",
              "error",
              "timeout",
              "interrupted"
            ],
            "title": "Status",
            "description": "The current status of the run."
          },
          "metadata": {
            "type": "object",
            "title": "Metadata",
            "description": "Custom metadata for the run."
          },
          "kwargs": { "type": "object", "title": "Kwargs" },
          "multitask_strategy": {
            "type": "string",
            "enum": ["reject", "enqueue", "rollback", "interrupt"],
            "title": "Multitask Strategy",
            "description": "Strategy for handling concurrent runs."
          }
        },
        "title": "Run"
      },
      "RunCreateStateful": {
        "type": "object",
        "required": ["assistant_id"],
        "properties": {
          "assistant_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "string" }
            ],
            "title": "Assistant ID",
            "description": "The assistant to use. Can be UUID or graph name."
          },
          "input": {
            "anyOf": [
              { "type": "object" },
              { "type": "array" },
              { "type": "string" },
              { "type": "integer" },
              { "type": "boolean" },
              { "type": "null" }
            ],
            "title": "Input",
            "description": "The input to the graph."
          },
          "command": {
            "type": "object",
            "title": "Command",
            "description": "Command to control graph execution (update, resume, goto)."
          },
          "checkpoint": {
            "type": "object",
            "title": "Checkpoint",
            "description": "Checkpoint to resume from."
          },
          "metadata": {
            "type": "object",
            "title": "Metadata",
            "description": "Metadata to attach to the run."
          },
          "config": {
            "type": "object",
            "title": "Config",
            "description": "Configuration for the graph."
          },
          "context": {
            "type": "object",
            "title": "Context",
            "description": "Context to pass to the graph."
          },
          "webhook": {
            "type": "string",
            "format": "uri",
            "title": "Webhook",
            "description": "Webhook URL to call on run completion."
          },
          "interrupt_before": {
            "anyOf": [
              { "type": "string", "enum": ["*"] },
              { "type": "array", "items": { "type": "string" } }
            ],
            "title": "Interrupt Before",
            "description": "Nodes to interrupt before execution."
          },
          "interrupt_after": {
            "anyOf": [
              { "type": "string", "enum": ["*"] },
              { "type": "array", "items": { "type": "string" } }
            ],
            "title": "Interrupt After",
            "description": "Nodes to interrupt after execution."
          },
          "stream_mode": {
            "anyOf": [
              {
                "type": "array",
                "items": {
                  "type": "string",
                  "enum": [
                    "values",
                    "updates",
                    "messages",
                    "messages-tuple",
                    "debug",
                    "events",
                    "custom"
                  ]
                }
              },
              {
                "type": "string",
                "enum": [
                  "values",
                  "updates",
                  "messages",
                  "messages-tuple",
                  "debug",
                  "events",
                  "custom"
                ]
              }
            ],
            "title": "Stream Mode",
            "description": "What to stream back.",
            "default": ["values"]
          },
          "stream_subgraphs": {
            "type": "boolean",
            "title": "Stream Subgraphs",
            "description": "Whether to stream subgraph events.",
            "default": false
          },
          "stream_resumable": {
            "type": "boolean",
            "title": "Stream Resumable",
            "description": "Whether to stream resumable checkpoints.",
            "default": false
          },
          "on_disconnect": {
            "type": "string",
            "enum": ["cancel", "continue"],
            "title": "On Disconnect",
            "description": "What to do when client disconnects.",
            "default": "continue"
          },
          "feedback_keys": {
            "type": "array",
            "items": { "type": "string" },
            "title": "Feedback Keys",
            "description": "Keys to collect feedback on."
          },
          "multitask_strategy": {
            "type": "string",
            "enum": ["reject", "enqueue", "rollback", "interrupt"],
            "title": "Multitask Strategy",
            "description": "How to handle concurrent runs on the same thread.",
            "default": "enqueue"
          },
          "if_not_exists": {
            "type": "string",
            "enum": ["create", "reject"],
            "title": "If Not Exists",
            "description": "What to do if the thread doesn't exist.",
            "default": "reject"
          },
          "after_seconds": {
            "type": "number",
            "title": "After Seconds",
            "description": "Delay before starting the run."
          },
          "checkpoint_during": {
            "type": "boolean",
            "title": "Checkpoint During",
            "description": "Whether to create checkpoints during execution.",
            "default": false
          },
          "durability": {
            "type": "string",
            "enum": ["sync", "async", "exit"],
            "title": "Durability",
            "description": "Durability mode for the run.",
            "default": "async"
          }
        },
        "title": "RunCreateStateful",
        "description": "Payload for creating a stateful run on a thread."
      },
...
    }
  }
}